+++ packages/frontend/core/src/blocksuite/ai/components/ai-item/types.ts
@@ -28,44 +28,3 @@ export interface AISubItemConfig {
   testId?: string;
   handler?: (host: EditorHost) => void;
 }
-
-abstract class BaseAIError extends Error {
-  abstract readonly type: AIErrorType;
-}
-
-export enum AIErrorType {
-  GeneralNetworkError = 'GeneralNetworkError',
-  PaymentRequired = 'PaymentRequired',
-  Unauthorized = 'Unauthorized',
-}
-
-export class UnauthorizedError extends BaseAIError {
-  readonly type = AIErrorType.Unauthorized;
-
-  constructor() {
-    super('Unauthorized');
-  }
-}
-
-// user has used up the quota
-export class PaymentRequiredError extends BaseAIError {
-  readonly type = AIErrorType.PaymentRequired;
-
-  constructor() {
-    super('Payment required');
-  }
-}
-
-// general 500x error
-export class GeneralNetworkError extends BaseAIError {
-  readonly type = AIErrorType.GeneralNetworkError;
-
-  constructor(message: string = 'Network error') {
-    super(message);
-  }
-}
-
-export type AIError =
-  | UnauthorizedError
-  | PaymentRequiredError
-  | GeneralNetworkError;
