+++ packages/frontend/core/src/blocksuite/ai/provider/index.ts
@@ -1,3 +1,4 @@
 export * from './ai-provider';
 export * from './copilot-client';
+export * from './error';
 export * from './setup-provider';
